fix(design-ops): harden parity pack evidence
